DataProcessor/requirements.txt
azure-functions
requests
azure-storage-blob 
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
gunicorn
django-filter